Relicense as Apache 2
Modify readme to state license

Also reword text to stop implying this was a derived work of
rosettacode.org, which was a consulted source but bears no significance
in the final result of this parser.

We use other data structures, different tokens, and in general, a more
rich interface. This should not be considered a derived work of Joel F.
Klein's work.